WIP Update BLUEPRINT data to include all datasets
- Stephanie Hickey started with a subset of these which she dubbed "FlowSorted.Blood.WGBS.BLUEPRINT" (see https://github.com/PeteHaitch/MethylSeqData/issues/2).
But when I looked into it there are many more potentially relevant datasets than the ones she initially selected.
So I think it makes sense to package up the whole dataset and make it available, potentially with a helper function to select subsets of the data.

rather than just Stephanie's selected